Create a Profile on sign up
allauth doesn't let us set the user model during sign up so we can
either override the create_user function or suck it up.
tl;dr sucked it up and found a better way.

Changed CompleteProfile to a create view that associates the Profile to
the User who just registered.